backend/requirements.txt
fastapi==0.104.1
joblib==1.3.2
lightgbm==4.1.0
numpy==1.25.2
optuna==3.4.0
pandas==2.0.3
pydantic==2.5.2
PyYAML==6.0.1
scikit_learn==1.3.1
transliterate==1.10.2
uvicorn==0.24.0.post1
python-multipart==0.0.6